Web API partially done, possibility to get all the Ads, a particular ad by Id and to delete an ad by id

+++ bea.test/dal/map/AdMapTest.cs
@@ -92,7 +92,7 @@ namespace Bea.Test.Dal.map
 
                 ad.AddImage(img);
                 repo.Save(ad);
-                int aId = ad.Id;
+                long aId = ad.Id;
 
                 repo.Flush();
                 repo.Clear();
